STCC-202 reimplemented metadata http retries
scratchwebapi.downloadProjectMetaData retries to build the connection
if the retry parameter is true.

It only retries if no connection to the server can be made,
or an unecpected statuscode is returned, but not when the server
answers with a 404 (then it logs it and returns None).

Removed old jsoup try/except in call of downloadProjectMetaData